Moves the beecoin shop to the DB

+++ code/modules/metacoin/metacoin.dm
@@ -21,6 +21,8 @@
 	inc_metabalance(METACOIN_TENMINUTELIVING_REWARD, FALSE)
 
 /client/proc/get_metabalance()
+	if(!isnull(metacoin_balance))
+		return metacoin_balance
 	var/datum/DBQuery/query_get_metacoins = SSdbcore.NewQuery(
 		"SELECT metacoins FROM [format_table_name("player")] WHERE ckey = :ckey",
 		list("ckey" = ckey)
@@ -31,27 +33,34 @@
 			mc_count = query_get_metacoins.item[1]
 
 	qdel(query_get_metacoins)
-	return text2num(mc_count)
+	metacoin_balance = text2num(mc_count)
+	return metacoin_balance
 
 /client/proc/set_metacoin_count(mc_count, ann=TRUE)
+	if(IsAdminAdvancedProcCall())
+		return
 	var/datum/DBQuery/query_set_metacoins = SSdbcore.NewQuery(
 		"UPDATE [format_table_name("player")] SET metacoins = :mc_count WHERE ckey = :ckey",
-		list("mc_count" = mc_count, "ckey" = ckey)	
+		list("mc_count" = mc_count, "ckey" = ckey)
 	)
 	query_set_metacoins.warn_execute()
 	qdel(query_set_metacoins)
+	metacoin_balance = mc_count
 	if(ann)
 		to_chat(src, "<span class='rose bold'>Your new metacoin balance is [mc_count]!</span>")
 
 /client/proc/inc_metabalance(mc_count, ann=TRUE, reason=null)
+	if(IsAdminAdvancedProcCall())
+		return
 	if(mc_count >= 0 && !CONFIG_GET(flag/grant_metacurrency))
 		return
 	var/datum/DBQuery/query_inc_metacoins = SSdbcore.NewQuery(
 		"UPDATE [format_table_name("player")] SET metacoins = metacoins + :mc_count WHERE ckey = :ckey",
-		list("mc_count" = mc_count, "ckey" = ckey)	
+		list("mc_count" = mc_count, "ckey" = ckey)
 	)
 	query_inc_metacoins.warn_execute()
 	qdel(query_inc_metacoins)
+	metacoin_balance += mc_count
 	if(ann)
 		if(reason)
 			to_chat(src, "<span class='rose bold'>[abs(mc_count)] [CONFIG_GET(string/metacurrency_name)]\s have been [mc_count >= 0 ? "deposited to" : "withdrawn from"] your account! Reason: [reason]</span>")
